Stop using source address as a filter; use TTL instead.
In order to avoid relay loops, the previous code only relayed the packet
if it had a source address that matched the default source address for
packets sent to the global broadcast address. If the packet had a
different source address, it would assume it came from WinIPBroadcast
itself and ignore it.

This works fine as long as the sending process leaves the socket unbound
to any source address while sending broadcast packets. However, if the
process explicitly binds the socket to a specific source address, and
that source address doesn't match the default source address, then
WinIPBroadcast will not work.

This commit fixes the issue by unconditionally relaying packets no
matter what their source address is. To avoid relay loops, we use a
different approach which consists in setting the TTL of outgoing packets
to just 1. When these packets come back in they will be dropped because
of their TTL. This is not a perfect solution, but it is arguably
superior to the previous approach.